Feat/person filter

+++ internal/domain/models/base_entity.go
@@ -2,8 +2,9 @@ package models
 
 import "time"
 
+// TODO: Improve filtering when it comes to dates.
 type BaseEntity struct {
-	Id        int        `json:"id"`
+	Id        int64      `json:"id"`
 	CreatedAt *time.Time `json:"created_at"`
 	UpdatedAt *time.Time `json:"updated_at"`
 	DeletedAt *time.Time `json:"deleted_at"`
