+++ internal/domain/models/person.go
@@ -2,7 +2,7 @@ package models
 
 type Person struct {
 	Fullname string `json:"fullname"`
-	Age      int8   `json:"age"`
+	Age      int64  `json:"age"`
 	Email    string `json:"email"`
 	BaseEntity
 }
